move setupProgramsAndCourses to its proper spot in test-utils
