+++ frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/test-utils.ts
@@ -6,6 +6,13 @@ import {
   EnrollmentStatus,
 } from "./types"
 import type { DashboardCourse } from "./types"
+import * as u from "api/test-utils"
+import { urls, factories } from "api/mitxonline-test-utils"
+import { setMockResponse } from "../../../test-utils"
+import { mockOrgData } from "api/mitxonline-hooks/enrollment"
+
+const makeCourses = factories.courses.courses
+const makeProgram = factories.programs.program
 
 const dashboardCourse: PartialFactory<DashboardCourse> = (...overrides) => {
   return mergeOverrides<DashboardCourse>(
@@ -31,4 +38,39 @@ const dashboardCourse: PartialFactory<DashboardCourse> = (...overrides) => {
   )
 }
 
-export { dashboardCourse }
+const setupProgramsAndCourses = () => {
+  const user = u.factories.user.user()
+  setMockResponse.get(u.urls.userMe.get(), user)
+
+  const orgId = mockOrgData.orgX.id
+  const coursesA = makeCourses({ count: 4 })
+  const coursesB = makeCourses({ count: 3 })
+  const programA = makeProgram({
+    courses: coursesA.results.map((c) => c.id),
+  })
+  const programB = makeProgram({
+    courses: coursesB.results.map((c) => c.id),
+  })
+
+  setMockResponse.get(
+    urls.programs.programsList({ orgId: mockOrgData.orgX.id }),
+    { results: [programA, programB] },
+  )
+  setMockResponse.get(urls.courses.coursesList({ id: programA.courses }), {
+    results: coursesA.results,
+  })
+  setMockResponse.get(urls.courses.coursesList({ id: programB.courses }), {
+    results: coursesB.results,
+  })
+
+  return {
+    orgId,
+    user,
+    programA,
+    programB,
+    coursesA: coursesA.results,
+    coursesB: coursesB.results,
+  }
+}
+
+export { dashboardCourse, setupProgramsAndCourses }
